solution-blueprints/report-generation-engine/src/requirements.txt
# Copyright © Advanced Micro Devices, Inc., or its affiliates.
#
# SPDX-License-Identifier: MIT

# Dependencies for AMD Report Generation Engine Blueprint
# Generated from pyproject.toml

aiohttp==3.13.2
fastapi==0.124.0
flask==3.1.2
langchain==1.1.2
langchain-community==0.4.1
langchain-core==1.1.2
langchain-openai==1.1.1
langsmith==0.4.56
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
requests==2.32.5
streamlit==1.52.1
tavily-python==0.7.14
uvicorn==0.38.0
